111: Frontend - Added web.config to fix routing.
Previously, we had added a web.config file to the project gui root
folder.  This didn't work.

After researching it further, according to what I've read, this file
actually needs to be placed in the public folder.  I've added the file
to the public folder so that we can try a deploy and see what happens.