111: Node POC - Setup to handle managed identity
Modified database connection setup to handle managed identity if
provided in environment variable.

By default we assume we are using managed identity and the config is
setup as such.

if AZURE_MANAGED_IDENTITY is not provided (empty string), and
MSSQL_PASS is provided (string), then it assumes it's using the
regular password MSSQL authentication.

If neither of the above environment variables are provided, but
DATABASE_MOCK is provided and set to 'true', then the database will be
mocked.

If none of the above variables are set, and/or if DATABASE_MOCK is
'false', then we will throw an error that no database auth was set.

In addition to those updates, we added proper handling for fetching all
cases and filtering via the chapter in our DB query, and handle SQL
injection properly.