111: Added login endpoint and updated /cases endpoint
We now have a login endpoint to supply a first name / last name and
receive back a user id or "professional code".

We are managing the user id on the front end and then passing it in our
call to the /cases endpoint, along with the chapter name.

/cases endpoint now returns filtered results for a particular user id
and chapter type.

We need to remember to update the user local data gateway and also
update the test cases.