CAMS-85: Fixed broken test
Thank you ChatGPT!

ChatGPT helped me solve an issue that came up when trying to mock a class
module where one property had to change but I still wanted to use the other
original properties and methods of the class.

The error I was getting was basically similar to
TypeError: foo_bar_1.FooBar is not a constructor.

Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>
Co-authored-by: Fritz Madden <fmadden@flexion.us>