Get count of cases assigned to attorneys
We were randomly generating the numbers displayed for how many cases are
assigned to an attorney when displaying the assignment modal. For the
cosmosdb implementation we now query the collection and return the
count. For the local implementation we are just returning an empty
array for now. We need to decide on how to implement this locally.

Coverage for the backend is now below 90% of lines and tests need to be
added.

[skip ci]

Jira ticket: CAMS-148