Tell Pa11y the correct url to test
With the addition of an id attribute to the html for ToggleModalButton
we now have a duplicate id which fails the test. If we don't actually
want id's we can just revert that change, but if we do want them, we
should ensure that each element has a unique id.

[skip ci]

Jira ticket: CAMS-47