Fix mock of api response
When I wrote the mock for api, I based it on how I mocked the attorneys
api, but that works differently. We were returning just case details
when we should have been returning an object with a body property which
had a value of another object with a caseDetails property.

Jira ticket: CAMS-164

Co-authored-by: Michael Ly <105381067+magile32@users.noreply.github.com>