Get closed date from transaction table
We get the list of closed dates for the case from the transaction table
but have not handled getting the latest date yet. Note that we have to
add 2000 to the year because just using '23' results in a date in 1923.
If dates prior to 2000 are expected we will have to find out how to
handle that.

Jira ticket: CAMS-167