Fix package-lock.json out of sync
We suspect that the npm version is what is causing the out-of-sync
issue that we sometimes experience when the package-lock.json file has
been updated. We also noticed that GitHub Actions is using a newer
version of Node than we have available in Azure, so we explicitly build
for 18.17.1 for now.

Jira ticket: CAMS-195

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Michael Ly <105381067+magile32@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>