Fixed font loading issues, optimized image load
Fonts were not being referenced correctly in the css.  I resolved this
by having the build copy the USWDS fonts over to the expected
directory.

In the process I noticed that the images could be loaded the same way,
and by making the switch, the image use for the sprite can be
simplified.  Therefore I also updated the way images are copied on
build.  The result is that all images from uswds are now included, as
well as all fonts.  This may be seen as a bad thing if we do not need to
utilize all of those images.  This can be refined later if need be.

Jira ticket: CAMS-197