Add a helper function for handling query success
We have some repetitive code which differs only in what steps to take
when the query of the DXTR database is successful. Writing tests to
cover the negative scenarios for each block of code was going to be
repetitive and verbose. So we create a function which takes a callback
to handle the success steps while keeping the testing of the else block
much cleaner. We only apply this to the parties query for now so there
is some refactoring to do for the other queries.

Jira ticket: CAMS-181

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Michael Ly <105381067+magile32@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>