WIP - Move assignment repository to factory
We need to remove the case assignment repository from the application
context and only get it from the factory. The problem with that is that
it was being treated as a singleton which will not work with the factory
 as it is currently implemented. Not treating the repository as a
singleton makes the most sense because it cannot persist across
invocations of the function app anyway. However, this causes failures
in the case assignment controller test which were relying on test
behavior that does not match actual behavior. All tests are passing but
the frontend is showing a 500 error even though the api is responding
with a 200.

Jira ticket: CAMS-141