Remove commented out tests
We had some tests that were relying on behavior that was a consequence
of treating the case assignment repository as if it persisted accross
api calls, which it never did. Removing the repository from the context
resulted in the tests failing because the repository is now instantiated
 for each call to the controller which is the same behavior as at
runtime. Several attempts to test exceptional behavior via mocks were
not successful. We will address in a DevEx task.

Jira ticket: CAMS-141

Co-authored-by: Arthur Morrow <133667008+amorrow-flexion@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>