Handle errors better when accessing Cosmos DB
If we fail to retrieve assignments when looking up attorneys and cases
we log the error and continue. But if we encounter an error when trying
to assign attorneys, we throw a useful error to let the user know that
we were either unable to create the assignment or unable to verify that
the case was not already assigned.

Jira ticket: CAMS-216

Co-authored-by: Arthur Morrow <133667008+amorrow-flexion@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>