Make UI improvements
Cherry picked changes from another branch that had some bad changes
merged into it.

The left nav content is known but was only being shown when isLoading is
false.
The docket record content was taking up too much space.
The mock data for use with pa11y was not mocking the docket record. The
backend mock was also not taking advantage of the new pattern we've
begun using to get the data from a constant in a neighboring TypeScript
file.

Jira ticket: CAMS-219

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Mrinalini Sinha <20514925+mrinsin@users.noreply.github.com>
co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>
Co-authored-by: Michael Ly <105381067+magile32@users.noreply.github.com>