Slight change to padding and placement of button
The button looks a little more symetrical and more like the button in
the other USTP app if it has less padding on left and right.  Also moved
the button a few pixes away from the right side because it was squished
up against the scroll bar.

Jira ticket: CAMS-238