Cleanup
When we run npm run clean for user-interface and backend, we now delete
the ./coverage directory. This is so that if we need to zip up the code
(for Veracode runs for example) we can just run the script and then do
the zip. We also remove some unnecessary logging.

Jira ticket: CAMS-324

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Arthur Morrow <133667008+amorrow-flexion@users.noreply.github.com>
Co-authored-by: Michael Ly <105381067+magile32@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>