Replace React fragment in tbody with tr
We needed two tr elements per case, but can't return two elements so we
were previously returning a React fragment (i.e. <></>) and this was the
offending element for the unique key error we were encountering in the
console. This is interesting to note because the React fragment is
often ignored as it is not an element that appears in the html in the
browser.

Jira ticket: CAMS-270

Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>