WIP: Creating E2E test for transfer orders
Currently working to create an e2e test in playwright which matches that
of Cypress. . Almost complete.  There are a couple issues we are having
remaining.

I can not seem to get past entering a case number in the new-case field.
Others got further, but failed in the confirmation that a new tab was
opened properly in the browser.

Jira ticket: CAMS-269