Tried to get date range picker test to pass.
There is a weird unexplainable behavior in vitest it seems.
if we console.log the value, we see the correct value.
The internal state changes properly with the debugger.
However, the assertion fails with an incorect return value

Left a TODO in the test to resolve at a later date.

*PUNT*!

Jira ticket: CAMS-348

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>,