separated suggested case test from manual case tests
The tests would not pass when coexisting in a single file.
It seems also that we were trying to use Chapter15MockApi while
simultaneously mocking Api.  That's not going to work.  We have to do
one or the other.

I tried to simply get rid of the older Chapter15MockApi use but that
only caused more problems.

Separating these tests into 2 files is a working temporary solution, but
it would be better to figure out what was causing the issue and resolve
it.

Coverage for the file is now above 90%.

Jira ticket: CAMS-338