UI update and 2 tests for errors now passing.
Updated the UI for the Transfer Order screen so that, when the search
for suggested cases comes up empty, the table is not displayed, the
alert is displayed, and the form is displayed.

This makes for an easier to understand screen when no suggestions are
found.  Updated wording for alert and "select new case" instructions to
work better with the new layout.

Jira ticket: CAMS-338

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>