big updates to consolidation accordion
Moved Consolidation Type and Lead Case selection to accordion and pulled
it out of the modal.

Modal is displaying correct information for approval.
Modal could use a loading state for attorney list, or perhaps pass
attorney list into it.
Modal Verify button is still disabled for some reason.

Lead case selection is working with the buttons in the table.

We still have to wire up form for "lead case not listed".

Still have styling and some work to do on the accordion.

Jira ticket: CAMS-390

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>