Added check for approve button
Approve button has an additional check now to ensure that if there is
only 1 child case, that that case is not also designated as the lead
case.  A consolidation must have at least 2 cases total.  If only 1
child case is available, then a second case must be the lead, by
entering it into the "case not listed" form.

Jira ticket: CAMS-396