Fixed issues with select box and random order generation
The random generation of orders was using a different division each time
and therefore generating a different number of filtered court offices.
That was creating an unreliable list in the select box.

Fixed by simply grabbing the first item in the list each time. (there is
always at least 1 entry)

Jira ticket: CAMS-396

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>