All desired functionality seems to be in place
ComboBox component looks to be working well with desired features.
For a list of requirements satisfied see docs/ui-components/Combobox.md

Tests can be writen according to the features listed in that markdown
file.

The component still needs to be plugged into the search screen
functionality to retrieve results.

Jira ticket: CAMS-375

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>
Co-authored-by: Sam Johnson <171059116+samjohnsonflexion@users.noreply.github.com>