Verify that bearer token possibly matches jwt spec
Veracode flagged the use of the bearer token as a violation of CWE-99.
This is probably not a problem for us, but it is possible that a
malformed or malicious authorization header could make it all the way
to us calling Okta with it. This change decodes the token to at least
verify that it is probably a valid jwt before we use it.

Jira ticket: CAMS-405

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>,
Co-authored-by: Sam Johnson <171059116+samjohnsonflexion@users.noreply.github.com>