Add simple local storage gateway
Using storage gateway for an externalized group mapping configuration
file. For now this file is a hard coded configuration returned by the
gateway. The use of the gateway interface provides us an abstraction
point whereby the configuration file can be stored in a storage service
at a later date.

[skip ci]

Jira ticket: CAMS-406