Can change Alert message for SearchResults
We added the ability to specify the Alert message for Search Results
when no results were returned.  This was done because we are using the
Search Results on both the Search screen and the My Cases screen.  We
need to have a different alert message on each screen.

Jira ticket: CAMS-48

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Sam Johnson <171059116+samjohnsonflexion@users.noreply.github.com>