Fixed and simplified modal for Accessibility fixes
Fixed Modal and Open-Modal-Button so that when Escape key is pressed,
the open modal button is focused properly.

While at it, I stripped out some code that is not used anywhere and
likely won't be used, and also changed the ToggleModalButton to an
OpenModalButton since the button will likely never toggle anything, but
rather just open a modal.  This is because it is not possible to click
on the toggle button while the modal is open.  Thus, the button will
never toggle shut -- it only opens the modal.

Simplified code and brought test coverage up to 100% on the
OpenModalButton and close to 98% on the Modal.

Jira ticket: CAMS-418