Accept a single argument for createAuditRecord
We want to be able to use zero, one, or both of two optional arguments.
We modify the function signature to accept a single object containing
the required `record` as well as the two situationally required
`session` and `override` arguments. We also document the appropriate use
with JSDocs.

Jira ticket: CAMS-282

Co-authored-by: Arthur Morrow <133667008+amorrow-flexion@users.noreply.github.com>