Fixed menu items not available when resized or zoomed
A diviated slightly from the USWDS header layout, because it did not
accomodate for moving the menu from the right to underneath the logo and
title and because to some extent, USWDS layout didn't make sense in
terms of navigation controls.

I also had to override a lot of CSS for the media size of width smaller
than 1024px.  This was because their usa-nav class hides the menu, sets
it up to display has a pull-down menu and adds all kinds of styling that
we do not want.  I however, felt it necessary to stick with the usa-nav
class for various reasons.

Jira ticket: CAMS-426