Fixed issue with screen reader reading alert
I added the ability to refresh the alert messages on screen after the
focus has been reestablished to the input field.  This causes the screen
reader to re-read the alert after the focus.  Basically, we add an
aria-hidden attribute to the alert and 100ms later remove it, which
causes the screen reader to detect that the alert has changed and
re-read it.

Jira ticket: CAMS-420