WIP - Use both id and _id to make finds work
There does not seem to be a way to query by _id with the node client.
So we create a UUID for id and query by that for replacing case
assignments. We will need to modify any other uses of replace or delete
that use the id for the query filter.

Jira ticket: CAMS-442

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Arthur Morrow <133667008+amorrow-flexion@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>,