Disable E2E test because bad test data.
The consolidaion E2E is disabled temporarily to allow
the CI/CD pipeline to continue. The reason the test is
disabled is because the pending order the test uses
now includes a case that is already cosolidated so the
verify button is never enabled allowing the test to
pass.

Jira ticket: CAMS-482

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>