Pin to old version of @types/mssql
The move from 9.1.5 to 9.1.6 brings in a change which cascades to a move
to a version of the tedious package which is causing our build to break.
For now, we have pinned the version of our dependency and provided an
override for @types/tedious. We should keep an eye on this to come
current at some point. Also of note, we are still pending an mssql
version upgrade.

Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>