Add sort fields to the index
We add a caseNumber fields to CaseBasics which may be used for searching
cases and any other logic that may benefit from operating on the caseId
without the division number. The default sort for the search screen for
example will involve both the dateFiled and caseNumber fields.

Jira ticket: CAMS-421