Updated the way form data is handled for casenotes
I removed the formData state variable from the component and added a
draftMode state variable.

The form data is managed completely by the Local Storage instead of the
component state and only the display of the alert is handled by local
state.  This greatly simplifies the logic and creates a cleaner
interface for data manipulation.

Jira ticket: CAMS-519