Completed first draft of CaseNoteModal
There are a couple tests I commented out that need to be moved to the
new test module for CaseNoteModal or deleted if not needed.

We are also lacking on coverage.  I just worked to get existing tests to
pass, and added a few for the modal.

Jira ticket: CAMS-527

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Arthur Morrow <133667008+amorrow-flexion@users.noreply.github.com>
Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>