WIP - Begin AddCaseModal
To support adding cases to consolidations we are converting the lead
case form into an add case modal. An unlisted child or lead case can be
added to the consolidation via the modal and then any case may be marked
as lead just as they can be today.

[skip ci]

Jira ticket: CAMS-548

Co-authored-by: Fritz Madden <96319835+fmaddenflx@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>
Co-authored-by: Nate Elliott <32004008+En-8@users.noreply.github.com>