Reworked tests in case-assignment.test.ts
Fixed mocking for each test.
New test is failing for the following reason:

The user so far, is found, but the office and division
codes assigned to this user does not match 081 which is the office we’re looking for
so this user is not a member of that office. case-assignment.ts line 44

Jira ticket: CAMS-552

Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>