WIP Updated sanitization on the backend
Still todo: We had removed the validateArchiveRequestParameters function
but in favor of a single validateRequestParameters function.

However, we may need to put it back in place because when validating
archive values, we don't want to care about title or content, but the
validation is going to throw when title or content are missing.

That however, IS a VALID test when testing POST and PUT content.

Jira ticket: CAMS-549

Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>
Co-authored-by: Nate Elliott <32004008+En-8@users.noreply.github.com>