Use a mode property instead of id for mode checks
We also do not save form state for edits as this interferes with the
create form state. If we need to save drafts for edits, we should do so
with the id of the existing note included in the cache and display an
alert with the case note rather than at the top.

Jira ticket: CAMS-526

Co-authored-by: Nate Elliott <32004008+En-8@users.noreply.github.com>