Properly define assignees during consolidation
We were improperly using the lead case id as the 'id' property for
assignees so no valid assignments were found as we were comparing case
id's to Okta user id's in the application of a filter. We have updated
the test cases to expect that the userId is what is passed in the id
field when calling the assignments use case during consolidation
handling.

Jira ticket: CAMS-579