Slice 1 almost complete.
Slice 1 is to add the ability to track the bold formatting of currently
selected text or the text inside which the cursor is placed. This
tracking should work if the cursor is collapsed and udpate as the
keyboard moves the cursor, or text is clicked on.

That much is complete and working.

To finish slice 1, we'll need to ensure that by selecting text, the
editor is able to detect and log the current bold format of the selected
text.

Jira ticket: CAMS-