Slice 1 fully implemented.
Highlighting text now correctly indicates if the text under highlight is
bold or not.  If the first part of the selection is not bold, then it
returns that selection is NOT bold.  If the first part of the selection
is bold, then it returns that it is bold.  In the future we will
indicated that if it's mixed formatting, that none of the formatting is
active.

Jira ticket: CAMS-526