WIP: Pa11y test fails
submit button for note can't be clicked until content is typed into
content editable field.  However, with pa11y we can't type directly into
the field. ChatGPT recommends we have a script run that enters text into
the field and another that fires the event that triggers the enabling of
the button. We should look for a better solution.

Jira ticket: CAMS-