Fixed formatting of contact info throughout
Created an EmailLink and FormattedAddress component and used it in
Trustee profile, audit history, case detail overview and anywhere it
seemed like it would fit.

Updated testing to have about 100% coverage or near 100% on all
EmailLink, FormattedAddress, or Trustee related tests and also coverage
for EmailLink and FormattedAddress on case detail screens and/or admin
screens that use them.

Jira ticket: CAMS-640

Co-authored-by: James Brooks <12275865+jamesobrooks@users.noreply.github.com>
Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>
Co-authored-by: Nate Elliott <32004008+En-8@users.noreply.github.com>