ComboBox now focuses on first item when tabbing
If the comboBox is open and user is typing, if there are characters in
the input, and the user presses tab key, then the first item in the list
gets focus.

If the user then presses enter to select the item, then the item becomes
selected and the combobox closes, and has focus.

If the user presses tab instead of enter, while the first item has
focus, then nothing is selected and the focus moves to the next input on
the screen.

This is the behavior demonstrated on USWDS

Jira ticket: CAMS-584

Co-authored-by: Brian Posey <15091170+btposey@users.noreply.github.com>