Simplified date validating for invalid dates.
Do to the fact that the html5 native date picker already does internal
validating, and entering an incomplete date, though the value in the
field looks like "02/dd/yyyy" for example, the actual value of the input
field is an empty string "".  So the only way to validate that a valid date
was entered correctly is to check input.validity.badInput for a value of
true.

This simplifies our code, but also limits the messages that we can
provide to the user.  Luckily, the only message we want to provide to
the user is "Must be a valid date mm/dd/yyyy."

For other errors like checking that the minimum date is reasonable
(after 1978), or checking that the date isn't more than 100 years into
the future, we'll need to first assume the user entered a valid date
(according to html5) and then check the value to ensure that it's not
out of range.

Jira ticket: CAMS-666